feat(urls): add endpoint for tree commits history
Added a new endpoint to fetch the commit history for a specific tree.
The new endpoint requires `origin`, `git_url`, and `git_branch` as
query parameters. It returns the earliest start time and build counts
for the commits.

TODO: Move this endpoint to a function so it can be called from the
tree details endpoint without making another request.