fix: builds filter
Before the changes, the treeCommitsHistory and treeDetailsView views were unable to handle more than one filter using the same field (and even if it worked it wouldn't be able to deal with where to put AND or OR). So I created get_grouped_filters to group all values ​​in an array into the same filter object. However, I received some errors in the query when dealing with a filter that only had one value in the array. Therefore, to avoid changing every query to always check whether it has one value or more, I decided to change the code in just one place, defining a string at first and changing it to an array when adding more values

Closes #438