feat(IssuesList): add a counter for unknown issues
Add a counter containing the number of failed tests with no linked
issues. To do it, add an optional parameter to `IssuesList` that
accept the number of total failed builds/boots/tests. If the parameter
is not present the behavior of the Issues List is equal as it was
before, otherwise it adds a new line to the list containing the number
of unknown issues.

Closes #544