fix(treeView): fix sql query to get correct status
Change the SQL query in the TreeView. Before It was considering only the
latest checkout when counting the status por builds, boots and tests
that was causing some status to be incorrect. This new query groups the
rows by the commit hash instead of the checkout ID so multiple checkouts
that correspond to the same tree now are used when counting the status,
returning the correct value

Closes #605