feat: consider .platform as hardware
Add a new logic to the backend where `environment_misc.platform` will
work as a fallback case for when the `environment_compatabile` field is
null in the database. When `environment_misc.plaftorm` is null, we'll
use `build.misc.platform`, but the logic of the Hardware filter will be
kept the same, meaning that the builds returned will be first the build
associated with the tests that have that hardware. This means that even
if `build.misc.platform` is null but the build is associated with
a test that have `environment_compatible` or `environment_misc.platform`
not null, it will also be added when filtering for that hardware

Closes #692