fix: add tree indexes state to header table
Added a new state `treeIndexesLength` that keep track of the size of
number of trees there are in the header table. This was necessary
because the backend expects a object that maps each tree index to its
respective selected commit, but it couldn't correctly get all the tree
indexes from the query parameter since it defaults to an empty array
when all trees are selected. This state is then used when the query
param is an empty array to create this object

Closes #783