:feat(listing-details): add and implements query inconsistency invalidator hook
add hook to cache bust in case of inconsistency between listing and
details

closes #833

Squashed stuff:

fix(listing-pages): move divs to more intuitive

Moved the divs containing the listing pages in Hardware.tsx and Trees.tsx
to the more intuitive position, the fixed layout is on the top, it make
senses that the divs also should be on the top

fix(tree-listing): update serializers and type

Removed total counts from TreeSerializer and TreeView that were not being used
handling. Added null_boots count to TreeView. # Please enter the commit message for your changes. Lines starting

Part of #833

```