fix(hardware-details): add correct join to query
When using the Django ORM, that does not support composite primary keys
as used by the Issues model, it would produce the incorrect join between
Issues and Incidents since it don't consider issue.version when
performing the join. To solve this, the django ORM query was replaced by
a raw query. That required some of the data to be treated in the
pydantic level since it was being fetched as string instead of dict