feat: copy local backend data to docker volume
Adds the behaviour to always copy new files to the backend-data volume when docker starts up; existing files won't be overwritten. This is made to keep a sync of local file additions, since the backend-data volume is mounted on top of the backend/data folder and makes local files inaccessible. Files that are in the local data are copied to a temporary folder, then the volume is mounted, the local files are synced to the volume files, and the temporary folder is deleted. Further changes to this behaviour should be done in the future.

Also changes the gunicorn initialization such that the notifications arguments don't go to it when running notifications command on docker

Closes #1222