fix: use default validators
When an endpoint is accessed without a parameter, the request.GET.get() will return None and pass it to the class. Since some value was passed, the basic 'default' won't be used (it's only used if there is no value being passed). In order to provide a centralized and working default value, validators are used in the query parameters, altering the value if it is None.

Closes #1340