feat(treedetails): allow lookup by commit tag or hash
Updated the TreeDetails endpoint and query logic to support resolving
either by commit hash or by tag name. This enables accessing the page via URLs
like /tree/{treeName}/{gitBranch}/{identifier}, where the identifier can be a hash or a tag.
In a case where the identifier reference multiple checkouts, the endpoints will return
the information related to the latest checkout with given information.

Closes #1320