refactor(models): revert foreign key renaming
Django assumes that the foreign keys are the name of the objects that is being related. It already places a '_id' at the end of the field. This also affects the readability of insertions, since they need to include the entire object of the relation, not just the id string