feat: add submission monitoring command
This command monitors a folder for new files, verifies the files, extracts the log_excerpt, adds its data to a queue for insertion and moves it to an archive subfolder. If there are any errors in the verification (such as invalid schema) it will be moved to a failed subfolder. The data is then removed from the queue and sent to be inserted via Django.

Closes #1369

Closes #1379