fix: timeout from nginx before gunicorn
Fixes a problem where the nginx would timeout way before our intended time of 240s. This time is now set to nginx (which returns error 504) instead of gunicorn (which returns error 500) for a better error message.

Also adds a message in the .env.example for the frontend in order to check if the frontend is bypassing nginx

Closes #1392