fix: reinforce required navigate params
Whenever we are navigating to a route with required path params (such as '/tree/{treeId}') we must pass the parameters. There were cases where the 'navigate' was going from one of those routes, but receiving the route via a variable and so the build step wasn't complaining, but we should reinforce those for correct code use and for future updates.