fix: repeated hardware commits
Checkouts were not being limited by origin, meaning that trees with checkouts in multiple origins were being returned multiple times. Since the hash is the only thing that matters and they are the same between origins, and that the hardware is visualized from a specific origin perspective, we can filter by origin and solve this problem.

Closes #1445