feat(ingester): build Django models concurrently
We're already in a thread pool, let's leverage that to build the Django
models (which involves some mappings) while other files are still
loading. Of course, this does not ensure that ALL mappings will be run
in parallel, as Python only runs Python code in one thread, but mapping
these in the DB thread would make it so *none* would.