feat(ingester): some improvements (#1588)
* fix: run `chmod +x` for script files

* fix(ingester): default value didn't match comment

* fix(ingester): rename `trees_name` to `tree_names`

Typo...

* chore(ingester): simplify sanitization calls, remove re-assignments

* feat(ingester): improve typing

* feat(ingester): keep track of `DirEntry` instead of `filename: str`

We were doing a bunch of operations to get info that we already had on
the DirEntry (but we were throwing that info away by not keeping it in
memory).

Besides, we were scanning the directory twice.

* feat(ingester): build Django models concurrently

We're already in a thread pool, let's leverage that to build the Django
models (which involves some mappings) while other files are still
loading. Of course, this does not ensure that ALL mappings will be run
in parallel, as Python only runs Python code in one thread, but mapping
these in the DB thread would make it so *none* would.